clean up utilities tests structure as well
